+++ android/src/main/kotlin/com/mapbox/maps/mapbox_maps/mapping/turf/TurfAdapters.kt
@@ -1,0 +1,14 @@
+package com.mapbox.maps.mapbox_maps.mapping.turf
+
+import com.google.gson.Gson
+import com.mapbox.geojson.Point
+
+fun Point.toList(): List<Any?> {
+  return listOf(mapOf("coordinates" to coordinates()))
+}
+
+object PointDecoder {
+  fun fromList(list: List<Any?>): Point {
+    return Point.fromJson(Gson().toJson(list.first()))
+  }
+}