+++ example/lib/full_map.dart
@@ -119,7 +119,7 @@ class FullMapState extends State<FullMap> {
                   coordinates: Position(
                 6.0033416748046875,
                 43.70908256335716,
-              )).toJson(),
+              )),
               zoom: 3.0),
           styleUri: MapboxStyles.LIGHT,
           textureView: true,
@@ -138,6 +138,7 @@ class FullMapState extends State<FullMap> {
           onStyleImageMissingListener: _onStyleImageMissingListener,
           onStyleImageUnusedListener: _onStyleImageUnusedListener,
           onResourceRequestListener: _onResourceRequestListener,
+          onLongTapListener: (coordinate) {},
         ));
   }
 }