+++ example/lib/polyline_annotations.dart
@@ -43,7 +43,7 @@ class PolylineAnnotationPageBodyState
       center:
         Point(coordinates: Position(
           0, 0
-        )).toJson(),
+        )),
       zoom: 1,
       pitch: 0)
     );