+++ lib/src/mapbox_map.dart
@@ -207,7 +207,7 @@ class MapboxMap extends ChangeNotifier {
 
   /// Convenience method that returns the `camera options` object for given parameters.
   Future<CameraOptions> cameraForCoordinatesPadding(
-    List<Map<String?, Object?>?> coordinates,
+    List<Point> coordinates,
     CameraOptions camera,
     MbxEdgeInsets? coordinatesPadding,
     double? maxZoom,
@@ -233,11 +233,8 @@ class MapboxMap extends ChangeNotifier {
           bounds, padding, bearing, pitch, maxZoom, offset);
 
   /// Convenience method that returns the `camera options` object for given parameters.
-  Future<CameraOptions> cameraForCoordinates(
-          List<Map<String?, Object?>?> coordinates,
-          MbxEdgeInsets padding,
-          double? bearing,
-          double? pitch) =>
+  Future<CameraOptions> cameraForCoordinates(List<Point> coordinates,
+          MbxEdgeInsets padding, double? bearing, double? pitch) =>
       _cameraManager.cameraForCoordinates(coordinates, padding, bearing, pitch);
 
   /// Convenience method that adjusts the provided `camera options` object for given parameters.
@@ -248,9 +245,7 @@ class MapboxMap extends ChangeNotifier {
   /// Note that this method may fail if the principal point of the projection is not inside the `box` or
   /// if there is no sufficient screen space, defined by principal point and the `box`, to fit the geometry.
   Future<CameraOptions> cameraForCoordinatesCameraOptions(
-          List<Map<String?, Object?>?> coordinates,
-          CameraOptions camera,
-          ScreenBox box) =>
+          List<Point> coordinates, CameraOptions camera, ScreenBox box) =>
       _cameraManager.cameraForCoordinatesCameraOptions(
           coordinates, camera, box);
 
@@ -288,8 +283,7 @@ class MapboxMap extends ChangeNotifier {
   ///
   /// The `screen coordinate` is in `logical pixels` relative to the top left corner
   /// of the map (not of the whole screen).
-  Future<ScreenCoordinate> pixelForCoordinate(
-          Map<String?, Object?> coordinate) =>
+  Future<ScreenCoordinate> pixelForCoordinate(Point coordinate) =>
       _cameraManager.pixelForCoordinate(coordinate);
 
   /// Calculates a geographical `coordinate` (i.e., longitude-latitude pair) that corresponds
@@ -297,7 +291,7 @@ class MapboxMap extends ChangeNotifier {
   ///
   /// The screen coordinate is in `logical pixels`relative to the top left corner
   /// of the map (not of the whole screen).
-  Future<Map<String?, Object?>> coordinateForPixel(ScreenCoordinate pixel) =>
+  Future<Point> coordinateForPixel(ScreenCoordinate pixel) =>
       _cameraManager.coordinateForPixel(pixel);
 
   /// Calculates `screen coordinates` that correspond to geographical `coordinates`
@@ -306,7 +300,7 @@ class MapboxMap extends ChangeNotifier {
   /// The `screen coordinates` are in `logical pixels` relative to the top left corner
   /// of the map (not of the whole screen).
   Future<List<ScreenCoordinate?>> pixelsForCoordinates(
-          List<Map<String?, Object?>?> coordinates) =>
+          List<Point> coordinates) =>
       _cameraManager.pixelsForCoordinates(coordinates);
 
   /// Calculates geographical `coordinates` (i.e., longitude-latitude pairs) that correspond
@@ -314,8 +308,7 @@ class MapboxMap extends ChangeNotifier {
   ///
   /// The screen coordinates are in `logical pixels` relative to the top left corner
   /// of the map (not of the whole screen).
-  Future<List<Map<String?, Object?>?>> coordinatesForPixels(
-          List<ScreenCoordinate?> pixels) =>
+  Future<List<Point?>> coordinatesForPixels(List<ScreenCoordinate?> pixels) =>
       _cameraManager.coordinatesForPixels(pixels);
 
   /// Changes the map view by any combination of center, zoom, bearing, and pitch, without an animated transition.
@@ -468,7 +461,7 @@ class MapboxMap extends ChangeNotifier {
 
   /// Gets elevation for the given coordinate.
   /// Note: Elevation is only available for the visible region on the screen and with terrain enabled.
-  Future<double?> getElevation(Map<String?, Object?> coordinate) =>
+  Future<double?> getElevation(Point coordinate) =>
       _mapInterface.getElevation(coordinate);
 
   /// Will load a new map style asynchronous from the specified URI.